refactor(fab): Consolidate DAG permissions code and modernize patterns - Replace hardcoded resource strings with constants in permissions_command.py   - Use RESOURCE_DAG_PREFIX and RESOURCE_DETAILS_MAP for DAG Run   - Remove deprecated Task Instance resource handling - Change to session handling with @provide_session decorator - Replace manual boolean parsing with airflow.utils.strings.to_boolean - Consolidate dag_permissions.py functionality into permissions_command.py - Consolidate test files for better maintainability
